Remove try-catch

+++ lidar_objects_tracker/include/lidar_objects_tracker/lmb_tracker.hpp
@@ -38,7 +38,7 @@ struct UpdateInfo
   };
 
   // Per track update info
-  std::map<size_t, TrackUpdateInfo> updates;
+  std::map<uint32_t, TrackUpdateInfo> updates;
 };
 
 /** @brief LMB Tracker for 2D point measurements
